Separated call reducer from user reducer, introdruced migration
