+++ src/store/reducers/call.reducer.js
@@ -4,16 +4,13 @@ import {updateObject} from "../../utils/utils";
 
 const reducer = (state = initialState, action) => {
   switch (action.type) {
-    case actionTypes.SET_AUTH_TOKEN:
+    case actionTypes.SET_INCOMING_CALL:
       return updateObject(state, action.payload);
-    case actionTypes.SET_USER_TYPE:
+    case actionTypes.END_CALL:
       return updateObject(state, action.payload);
-    case actionTypes.SET_INITIAL_LOGIN_OFF:
+    case actionTypes.SET_CALL_ACTIVE:
       return updateObject(state, action.payload);
-    case actionTypes.GENERIC_USER_FIELD_SET:
-      return updateObject(state, action.payload);
-    case actionTypes.RESET_USER:
-      return initialState;
+
     default:
       return state;
   }
