+++ src/store/reducers/root.reducer.js
@@ -1,10 +1,12 @@
-import { combineReducers } from "redux";
+import {combineReducers} from "redux";
 import AsyncStorage from "@react-native-community/async-storage";
-import { persistReducer } from "redux-persist";
+import {persistReducer, createMigrate} from "redux-persist";
 
 import auth from "./auth.reducer";
 import user from "./user.reducer";
+import call from "./call.reducer";
 import app from "./app.reducer";
+import {callState} from "./initialState";
 
 
 const authPersistConfig = {
@@ -12,17 +14,29 @@ const authPersistConfig = {
   storage: AsyncStorage
 };
 
+const migrations = {
+  0: (state) => {
+    return {
+      ...state,
+      call:callState
+    }
+  }
+}
 const rootPersistConfig = {
   key: "root",
-  version: 1,
+  version: 2,
   storage: AsyncStorage,
   blacklist: ["auth"],
+  migrate: createMigrate(migrations, {debug: true})
 };
 
+
+
 const rootReducer = combineReducers({
-  auth:persistReducer(authPersistConfig,auth),
+  auth: persistReducer(authPersistConfig, auth),
   user,
   app,
+  call
 });
 
 export default persistReducer(rootPersistConfig, rootReducer);
